fixed data utils and tested it, as well as tweaks to others

+++ Needles/src/Core/Structures.h
@@ -8,7 +8,7 @@ struct Seed {
 
 struct Data {
 	uint8_t* Bytes = nullptr;
-	uint8_t Size = 0;
+	size_t Size = 0;
 	bool Encrypted = false;
 #ifdef OPENSSL_READY
 	uint8_t* IV = nullptr;
